refactor: update dependancies after renaming file
